Add standard flake8 and isort configurations
These come from the cfpb/development repository.

The issue of different results between @cwdavies and me was that my govdelivery repo was checked out in a subfolder of cfgov-refresh, and thus it was inheriting the cfgov-refresh `.isort.cfg`. His was not. These changes give the govdelivery package its own explicit config.